Move remaining legacy test bases out of `tests.__init__`
