+++ tests/lib/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) 2015 The Pycroft Authors. See the AUTHORS file.
 # This file is part of the Pycroft project and licensed under the terms of
 # the Apache License, Version 2.0. See the LICENSE file for details.
-from tests import setup, teardown
+from tests.legacy_base import setup, teardown
