+++ tests/lib/user/test_move.py
@@ -5,7 +5,8 @@ from pycroft.model import session
 from pycroft.model.facilities import Room
 from pycroft.model.task import Task, UserTask, TaskStatus, TaskType
 from pycroft.model.task_serialization import UserMoveParams
-from tests import FactoryDataTestBase, factories, UserFactory, FactoryWithConfigDataTestBase
+from tests import factories, UserFactory
+from tests.legacy_base import FactoryDataTestBase, FactoryWithConfigDataTestBase
 from tests.factories import UserWithHostFactory, RoomFactory
 from tests.lib.user.task_helpers import create_task_and_execute
